Refactor component imports and update type paths for consistency
