+++ src/components/Thought/Visualizer/blocks/TextBlock.tsx
